Change external_code generator test to use float instead of int32_t because it does not require including a header file to get the right type for int32_t and that include is very difficult to do reliably when possibly compiling with clang that isn't fully installed, etc.

+++ test/generator/external_code_extern.cpp
@@ -1,5 +1,3 @@
-#include "HalideRuntime.h"
-
-extern "C" int32_t gen_extern_tester(int32_t in) {
+extern "C" float gen_extern_tester(float in) {
     return in + 42;
 }
